+++ tests/Unit/Cli/Routing/Provider/ServiceProviderTest.php
@@ -29,8 +29,8 @@ use Valkyrja\Cli\Routing\Collector\AttributeCollector;
 use Valkyrja\Cli\Routing\Collector\Contract\Collector;
 use Valkyrja\Cli\Routing\Contract\Router as RouterContract;
 use Valkyrja\Cli\Routing\Data;
+use Valkyrja\Cli\Routing\Dispatcher\Router;
 use Valkyrja\Cli\Routing\Provider\ServiceProvider;
-use Valkyrja\Cli\Routing\Router;
 use Valkyrja\Dispatcher\Contract\Dispatcher;
 use Valkyrja\Dispatcher\Data\MethodDispatch as DefaultDispatch;
 use Valkyrja\Reflection\Contract\Reflection;
